+++ packages/data-loader/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @fustilio/data-loader
 
+## 0.3.1
+
+### Patch Changes
+
+- remove wordnet specific implementations
+
 ## 0.3.0
 
 ### Minor Changes